pyproject.toml
[tool.poetry]
name = "transishow"
version = "0.1.0"
description = ""
authors = ["pysan3 <pysan3@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
rich = "^12.6.0"
fastapi = "^0.87.0"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
gunicorn = "^20.1.0"
sqlalchemy = "^1.4.44"
deta = "^1.1.0"
pyjwt = "^2.6.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
alembic = "^1.8.1"
python-slugify = "^7.0.0"
psycopg2 = "^2.9.5"
fastapi-mail = "^1.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
